[Website] Remove global Blueprint draft autosaves
