+++ packages/playground/personal-wp/src/lib/state/url/resolve-blueprint-from-url.ts
@@ -9,11 +9,9 @@ import {
 	isBlueprintBundle,
 	resolveRemoteBlueprint,
 } from '@wp-playground/client';
-import { OpfsFilesystemBackend } from '@wp-playground/storage';
 import { parseBlueprint } from './router';
 import { OverlayFilesystem, InMemoryFilesystem } from '@wp-playground/storage';
 import { RecommendedPHPVersion } from '@wp-playground/common';
-import { logger } from '@php-wasm/logger';
 import { decodeBlueprintHash } from './decode-blueprint-hash';
 
 export { decodeBlueprintHash };
@@ -26,9 +24,6 @@ export type BlueprintSource =
 	| {
 			type: 'personal-blueprint';
 			url: string;
-	  }
-	| {
-			type: 'last-autosave';
 	  }
 	| {
 			type: 'inline-string';
@@ -104,25 +99,6 @@ export async function resolveBlueprintFromURL(
 				type: 'remote-url',
 				url: blueprintUrl,
 			},
-		};
-	} else if (fragment === 'last-autosave') {
-		let bundle = undefined;
-		try {
-			bundle = await OpfsFilesystemBackend.fromPath(
-				'blueprints/last-edited-bundle',
-				true
-			);
-		} catch (error) {
-			logger.error(
-				'Failed to load the last edited blueprint from OPFS',
-				error
-			);
-		}
-		return {
-			blueprint:
-				bundle ||
-				((await resolveRemoteBlueprint(url.href)) as BlueprintV1),
-			source: { type: 'last-autosave' },
 		};
 	} else if (fragment.length) {
 		/*
