+++ packages/playground/website/src/components/blueprint-editor/blueprint-bundle-editor.module.css
@@ -151,51 +151,6 @@
 	display: none;
 }
 
-.autosaveOverlay {
-	position: absolute;
-	inset: 0;
-	background: rgba(0, 0, 0, 0.3);
-	backdrop-filter: blur(5px);
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	padding: 24px;
-	z-index: 6;
-}
-
-.autosaveCard {
-	width: min(440px, 90%);
-	background: #ffffff;
-	border-radius: 8px;
-	padding: 24px 28px;
-	box-shadow: 0 24px 48px rgba(0, 0, 0, 0.25);
-	display: flex;
-	flex-direction: column;
-	gap: 12px;
-}
-
-.autosaveTitle {
-	margin: 0;
-	font-size: 18px;
-}
-
-.autosaveMessage {
-	margin: 0;
-	color: var(--color-gray-700);
-	line-height: 1.5;
-}
-
-.autosaveActions {
-	display: flex;
-	flex-wrap: wrap;
-	gap: 10px;
-	margin-top: 6px;
-}
-
-.autosaveError :global(.components-notice) {
-	margin-top: 4px;
-}
-
 @media (max-width: 960px) {
 	.sidebarWrapper {
 		position: absolute;
