+++ packages/playground/website/src/lib/state/url/resolve-blueprint-from-url.ts
@@ -9,10 +9,8 @@ import {
 	isBlueprintBundle,
 	resolveRemoteBlueprint,
 } from '@wp-playground/client';
-import { OpfsFilesystemBackend } from '@wp-playground/storage';
 import { parseBlueprint, isMcpServerEnabled } from './router';
 import { OverlayFilesystem, InMemoryFilesystem } from '@wp-playground/storage';
-import { logger } from '@php-wasm/logger';
 import { decodeBlueprintHash } from './decode-blueprint-hash';
 import { getDefaultPhpVersionForWordPress } from '../../wordpress-version-compatibility';
 
@@ -22,9 +20,6 @@ export type BlueprintSource =
 	| {
 			type: 'remote-url';
 			url: string;
-	  }
-	| {
-			type: 'last-autosave';
 	  }
 	| {
 			type: 'inline-string';
@@ -108,25 +103,6 @@ export async function resolveBlueprintFromURL(
 				type: 'remote-url',
 				url: blueprintUrl,
 			},
-		};
-	} else if (fragment === 'last-autosave') {
-		let bundle = undefined;
-		try {
-			bundle = await OpfsFilesystemBackend.fromPath(
-				'blueprints/last-edited-bundle',
-				true
-			);
-		} catch (error) {
-			logger.error(
-				'Failed to load the last edited blueprint from OPFS',
-				error
-			);
-		}
-		return {
-			blueprint:
-				bundle ||
-				((await resolveRemoteBlueprint(url.href)) as BlueprintV1),
-			source: { type: 'last-autosave' },
 		};
 	} else if (fragment.length) {
 		if (isMcpServerEnabled()) {
